fix(lens): single-source the SendPlan interpreter in sendExec.ts
Extract the write-loop + grace + submit-Enter sequencing into executeSendPlan (an effectful leaf taking an injected write sink); cmdSend and the acceptance test both call it, so the test exercises the shipped path instead of a replica.

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
